feat(Deployment): Support for additional Deployment operations
Fixes https://github.com/godaddy/kubernetes-client/issues/47
