+++ lib/extensions.js
@@ -1,12 +1,12 @@
 'use strict';
 
 const ApiGroup = require('./api-group');
+const Deployments = require('./deployments');
 
 class Extensions extends ApiGroup {
   constructor(options) {
-    const resources = [
+    const commonResources = [
       'daemonsets',
-      'deployments',
       'horizontalpodautoscalers',
       'ingresses',
       'jobs',
@@ -16,8 +16,14 @@ class Extensions extends ApiGroup {
     options = Object.assign({}, options, {
       path: 'apis/extensions',
       version: options.version || 'v1beta1',
-      groupResources: resources,
-      namespaceResources: resources
+      groupResources: commonResources.concat(['deployments']),
+      //
+      // The custom Deployments objects implement functionality available only
+      // in namespaces.
+      //
+      namespaceResources: commonResources.concat([
+        { name: 'deployments', Constructor: Deployments }
+      ])
     });
     super(options);
   }
