+++ test/extensions.test.js
@@ -9,92 +9,6 @@ const beforeTesting = common.beforeTesting;
 const resourceName = 'test';
 
 describe('lib.extensions', () => {
-  describe('.deployments', () => {
-    const path = `/apis/extensions/v1beta1/namespaces/${ common.currentName }/deployments`;
-    const resourcePath = `${ path }/${ resourceName }`;
-    const deploymentObj = {
-      kind: 'Deployment',
-      metadata: {
-        name: resourceName
-      },
-      spec: {
-        selector: [
-          { app: 'test' }
-        ],
-        template: {
-          metadata: {
-            labels: {
-              app: 'test'
-            }
-          },
-          spec: {
-            containers: [{
-              name: 'test',
-              imagePullPolicy: 'Never',
-              image: 'test.com:5000/test:v0'
-            }]
-          }
-        }
-      }
-    };
-
-    beforeTesting('int', common.changeName);
-    beforeTesting('unit', () => {
-      const mockDeployment = {
-        kind: 'Deployment',
-        metadata: {
-          name: resourceName
-        }
-      };
-      nock(common.extensions.url)
-        .post(path)
-        .reply(201, mockDeployment)
-        .get(resourcePath)
-        .reply(200, mockDeployment)
-        .delete(resourcePath)
-        .reply(200, mockDeployment);
-    });
-
-    it('POSTs, GETs, and DELETEs', done => {
-      async.series([
-        next => {
-          common.extensions.ns.deployments.post({ body: deploymentObj }, next);
-        },
-        next => common.extensions.ns.deployments.get(resourceName, next),
-        next => common.extensions.ns.deployments.delete(resourceName, next)
-      ], (err, results) => {
-        assume(err).is.falsy();
-        const deployments = results[0];
-        assume(deployments.metadata.name).is.equal(resourceName);
-        done();
-      });
-    });
-
-    describe('lists', () => {
-      beforeTesting('int', common.changeName);
-      beforeTesting('unit', () => {
-        const mockDeploymentList = {
-          kind: 'DeploymentList',
-          items: []
-        };
-        nock(common.extensions.url)
-          .get(path)
-          .reply(200, mockDeploymentList);
-      });
-
-      it('returns DeploymentList', done => {
-        async.series([
-          next => common.extensions.ns.deployments.get(next)
-        ], (err, results) => {
-          assume(err).is.falsy();
-          const deploymentList = results[0];
-          assume(deploymentList.kind).is.equal('DeploymentList');
-          done();
-        });
-      });
-    });
-  });
-
   describe('.ds', () => {
     const path = `/apis/extensions/v1beta1/namespaces/${ common.currentName }/daemonsets`;
     const resourcePath = `${ path }/${ resourceName }`;
